+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [0.6.2]
+
+### Added
+
+* end to end testing around ssh collection
+
+### Changed
+
+* Invalid or missing ddc.yaml will now stop execution
+* earlier exit of remote execution of ddc if copying of ddc or ddc.yaml fails
+
 ## [0.6.1]
 
 ### Added
@@ -320,6 +331,7 @@
 
 - able to capture logs, configuration and diagnostic data from dremio clusters deployed on Kubernetes and on-prem
 
+[0.6.2]: https://github.com/dremio/dremio-diagnostic-collector/compare/v0.6.1...v0.6.2
 [0.6.1]: https://github.com/dremio/dremio-diagnostic-collector/compare/v0.6.0...v0.6.1
 [0.6.0]: https://github.com/dremio/dremio-diagnostic-collector/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/dremio/dremio-diagnostic-collector/compare/v0.4.0...v0.5.0